handle errors

+++ services/product_service.py
@@ -1,5 +1,6 @@
 from dataclasses import dataclass
 
+from exceptions import ProductNotFoundException, ProductAlreadyExistsException
 from repositories import ProductCache, ProductRepository
 from schemas import ProductSchema, CreateProductSchema, UpdtaeProductSchema
 
@@ -52,13 +53,26 @@ class ProductService:
         return product_schema
 
     def create_product(self, body: CreateProductSchema) -> ProductSchema:
+        check_product = self.product_repository.read_product_by_name(body.name)
+        if check_product:
+            raise ProductAlreadyExistsException
+
         product_id = self.product_repository.create_product(body)
         new_product = self.product_repository.read_product_by_id(product_id)
         return ProductSchema.model_validate(new_product)
 
     def update_product_name(self, product_id: int, product_name: str) -> UpdtaeProductSchema:
+        check_product = self.product_repository.read_product_by_id(product_id)
+
+        if check_product is None:
+            raise ProductNotFoundException
+
         updated_product = self.product_repository.update_product_name(product_id, product_name)
         return UpdtaeProductSchema.model_validate(updated_product)
 
     def delete_product(self, product_id: int) -> None:
+        check_product = self.product_repository.read_product_by_id(product_id)
+
+        if check_product is None:
+            raise ProductNotFoundException
         self.product_repository.delete_product(product_id)
